requirements.txt
aiosmtplib==4.0.1
asgiref==3.8.1
astroid==3.3.10
dill==0.4.0
Django==5.2.1
isort==6.0.1
mccabe==0.7.0
platformdirs==4.3.8
pylint==3.3.7
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
python-dotenv==1.1.0
sqlparse==0.5.3
tomlkit==0.13.2
